function that returns the value of a bit at a given index.

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -8,7 +8,7 @@
  */
 int get_bit(unsigned long int n, unsigned int index)
 {
-	unsigned long int count;
+	unsigned int count;
 
 	count = (n >> index) & 1;
